Get choice function
I added the menu, which works fine except for the quit case. I would have done the function not void where it could return the choice each time, but it is only suppose to display the menu at the beginning or when the user presses 'm'. If yall have any suggestions please lemme know and I'll fix it!